+++ build.gradle
@@ -95,7 +95,9 @@ allprojects {
   }
 
   repositories {
+    mavenLocal()
     mavenCentral()
+    jcenter()
     maven { url 'https://jitpack.io' }
     maven { url 'https://repo.spring.io/libs-milestone' }
   }
