+++ src/main/java/uk/gov/hmcts/reform/prl/controllers/hearingmanagement/HearingsManagementController.java
@@ -19,23 +19,13 @@ import org.springframework.web.bind.annotation.RequestHeader;
 import org.springframework.web.bind.annotation.RestController;
 import uk.gov.hmcts.reform.ccd.client.CoreCaseDataApi;
 import uk.gov.hmcts.reform.ccd.client.model.CallbackResponse;
-import uk.gov.hmcts.reform.ccd.client.model.CaseDataContent;
 import uk.gov.hmcts.reform.ccd.client.model.CaseDetails;
-import uk.gov.hmcts.reform.ccd.client.model.Event;
-import uk.gov.hmcts.reform.ccd.client.model.StartEventResponse;
 import uk.gov.hmcts.reform.idam.client.IdamClient;
-import uk.gov.hmcts.reform.prl.enums.LanguagePreference;
 import uk.gov.hmcts.reform.prl.enums.State;
-import uk.gov.hmcts.reform.prl.enums.YesOrNo;
 import uk.gov.hmcts.reform.prl.exception.HearingManagementValidationException;
-import uk.gov.hmcts.reform.prl.models.Element;
-import uk.gov.hmcts.reform.prl.models.complextypes.PartyDetails;
 import uk.gov.hmcts.reform.prl.models.dto.ccd.CaseData;
 import uk.gov.hmcts.reform.prl.models.dto.citizen.HearingDetailsRequest;
 import uk.gov.hmcts.reform.prl.models.dto.hearingmanagement.HearingRequest;
-import uk.gov.hmcts.reform.prl.models.dto.notify.EmailTemplateVars;
-import uk.gov.hmcts.reform.prl.models.dto.notify.HearingDetailsEmail;
-import uk.gov.hmcts.reform.prl.models.email.EmailTemplateNames;
 import uk.gov.hmcts.reform.prl.services.AuthorisationService;
 import uk.gov.hmcts.reform.prl.services.EmailService;
 import uk.gov.hmcts.reform.prl.services.SystemUserService;
@@ -43,14 +33,9 @@ import uk.gov.hmcts.reform.prl.services.citizen.CaseService;
 import uk.gov.hmcts.reform.prl.services.hearingmanagement.HearingManagementService;
 import uk.gov.hmcts.reform.prl.utils.CaseUtils;
 
-import java.util.Optional;
-
 import static javax.ws.rs.core.MediaType.APPLICATION_JSON;
 import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
 import static org.springframework.http.MediaType.MULTIPART_FORM_DATA_VALUE;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.C100_CASE_TYPE;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.CASE_TYPE;
-import static uk.gov.hmcts.reform.prl.constants.PrlAppsConstants.JURISDICTION;
 
 @Slf4j
 @RestController
@@ -93,6 +78,7 @@ public class HearingsManagementController {
     public void hearingManagementStateUpdate(@RequestHeader("serviceAuthorization") String s2sToken,
                                      @RequestBody HearingRequest hearingRequest) throws Exception {
 
+        log.info("");
         if (Boolean.FALSE.equals(authorisationService.authoriseService(s2sToken))) {
             throw new HearingManagementValidationException("Provide a valid s2s token");
         } else {
@@ -120,90 +106,10 @@ public class HearingsManagementController {
             .state(State.DECISION_OUTCOME)
             .build();
 
-        StartEventResponse startEventResponse =
-            coreCaseDataApi.startEventForCitizen(
-                authorisation,
-                s2sToken,
-                idamClient.getUserInfo(authorisation).getUid(),
-                JURISDICTION,
-                CASE_TYPE,
-                caseId,
-                eventId
-            );
-
-        CaseDataContent caseDataContent = CaseDataContent.builder()
-            .eventToken(startEventResponse.getToken())
-            .event(Event.builder()
-                       .id(startEventResponse.getEventId())
-                       .build())
-            .data(caseData).build();
-
-        coreCaseDataApi.submitEventForCitizen(
-            authorisation,
-            s2sToken,
-            idamClient.getUserInfo(authorisation).getUid(),
-            JURISDICTION,
-            CASE_TYPE,
-            caseId,
-            true,
-            caseDataContent
-        );
-        sendHearingDetailsEmailToCitizen(hearingDetailsRequest, caseData);
+        hearingManagementService.sendHearingDetailsEmailToCitizen(hearingDetailsRequest, caseData);
 
         return "SUCCESS";
     }
 
-    private void sendHearingDetailsEmailToCitizen(HearingDetailsRequest hearingDetailsRequest, CaseData caseData) {
 
-        String email = "";
-        String  partyId = hearingDetailsRequest.getPartyId();
-        if (caseData.getCaseTypeOfApplication().equals(C100_CASE_TYPE)) {
-            Optional<PartyDetails> applicantPartyDetails = caseData.getApplicants()
-
-                .stream()
-                .filter(applicant -> applicant.getId().equals(partyId))
-                .map(Element::getValue)
-                .findFirst();
-            if (applicantPartyDetails.isPresent()) {
-                if (applicantPartyDetails.get().getCanYouProvideEmailAddress().equals(YesOrNo.Yes)) {
-                    email = applicantPartyDetails.get().getEmail();
-                }
-            }
-
-            Optional<PartyDetails> respondenytPartyDetails = caseData.getRespondents()
-                .stream()
-                .filter(respondent -> respondent.getId().equals(partyId))
-                .map(Element::getValue)
-                .findFirst();
-
-            if (respondenytPartyDetails.isPresent() && email.isEmpty()) {
-                if (respondenytPartyDetails.get().getCanYouProvideEmailAddress().equals(YesOrNo.Yes)) {
-                    email = respondenytPartyDetails.get().getEmail();
-                }
-            }
-        } else {
-            if (caseData.getApplicantsFL401().getCanYouProvideEmailAddress().equals(YesOrNo.Yes)) {
-                email = caseData.getApplicantsFL401().getEmail();
-            }
-            if (caseData.getRespondentsFL401().getCanYouProvideEmailAddress().equals(YesOrNo.Yes) && email.isEmpty()) {
-                email = caseData.getRespondentsFL401().getEmail();
-            }
-        }
-
-        emailService.send(
-            "test@example.com",
-            EmailTemplateNames.HEARING_DETAILS,
-            buildHearingDateEmail(hearingDetailsRequest, caseData),
-            LanguagePreference.english
-        );
-    }
-
-    private EmailTemplateVars buildHearingDateEmail(HearingDetailsRequest hearingDetailsRequest, CaseData caseData) {
-        return HearingDetailsEmail.builder()
-            .caseReference(hearingDetailsRequest.getCaseId())
-            .caseName(caseData.getApplicantCaseName())
-            .partyName(hearingDetailsRequest.getPartyName())
-            .hearingDetailsPageLink(hearingDetailsUrl)
-            .build();
-    }
 }
